fix: extract frontend framework to its own boilerplates
