+++ boilerplates/plain-sentry/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": [
+    "../tsconfig.base.json"
+  ],
+  "compilerOptions": {
+    "types": ["vite/client", "@batijs/core/types"],
+    "lib": ["DOM", "DOM.Iterable", "ES2022"]
+  }
+}
