+++ boilerplates/react-sentry/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@batijs/sentry",
+  "name": "@batijs/react-sentry",
   "private": true,
   "version": "0.0.1",
   "description": "",
@@ -13,19 +13,12 @@
   "license": "MIT",
   "devDependencies": {
     "@batijs/compile": "workspace:*",
-    "@sentry/browser": "^8.27.0",
-    "@sentry/react": "^8.27.0",
-    "@sentry/solid": "^8.27.0",
-    "@sentry/vite-plugin": "^2.22.3",
-    "@sentry/vue": "^8.27.0",
-    "@types/node": "^18.19.44",
-    "dotenv": "^16.4.5",
-    "react": "^18.3.1",
+    "@types/node": "^18.19.14",
     "@types/react": "^18.3.4",
     "@types/react-dom": "^18.3.0",
-    "solid-js": "^1.8.21",
-    "vite": "^5.4.2",
-    "vue": "^3.4.38"
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "vite": "^5.4.2"
   },
   "dependencies": {
     "@batijs/core": "workspace:*"
@@ -35,18 +28,23 @@
   ],
   "bati": {
     "if": {
-      "flag": "sentry"
+      "flag": {
+        "$all": [
+          "react",
+          "sentry"
+        ]
+      }
     }
   },
   "exports": {
-    "./sentry.browser.config": {
-      "types": "./dist/types/sentry.browser.config.d.ts"
+    "./pages/sentry/+Page": {
+      "types": "./dist/types/pages/sentry/+Page.d.ts"
     }
   },
   "typesVersions": {
     "*": {
-      "sentry.browser.config": [
-        "./dist/types/sentry.browser.config.d.ts"
+      "pages/sentry/+Page": [
+        "./dist/types/pages/sentry/+Page.d.ts"
       ]
     }
   }
