--- boilerplates/sentry/files/pages/+client.ts
@@ -1,3 +1,0 @@
-/*{ @if (!it.BATI.has("vue")) }*/
-import "../sentry.browser.config";
-/*{ /if}*/
