+++ boilerplates/solid-sentry/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": [
+    "../tsconfig.base.json"
+  ],
+  "compilerOptions": {
+    "types": ["vite/client", "@types/node", "vike-solid/client", "@batijs/core/types"],
+    "jsx": "preserve",
+    "jsxImportSource": "solid-js",
+    "lib": ["DOM", "DOM.Iterable", "ES2022"]
+  }
+}
