+++ boilerplates/vue-sentry/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": ["../tsconfig.base.json"],
+  "compilerOptions": {
+    "types": ["vite/client", "@types/node", "@batijs/core/types"],
+    "jsx": "preserve",
+    "jsxImportSource": "vue",
+    "lib": ["DOM", "DOM.Iterable", "ES2022"]
+  }
+}
