+++ packages/cli/turbo.json
@@ -25,11 +25,13 @@
         "@batijs/hono#build",
         "@batijs/lucia-auth#build",
         "@batijs/mantine#build",
+        "@batijs/plain-sentry#build",
         "@batijs/prettier#build",
         "@batijs/prisma#build",
         "@batijs/react#build",
         "@batijs/react-firebase-auth#build",
         "@batijs/react-lucia-auth#build",
+        "@batijs/react-sentry#build",
         "@batijs/sentry#build",
         "@batijs/shared#build",
         "@batijs/shared-no-db#build",
@@ -39,6 +41,7 @@
         "@batijs/solid#build",
         "@batijs/solid-firebase-auth#build",
         "@batijs/solid-lucia-auth#build",
+        "@batijs/solid-sentry#build",
         "@batijs/sqlite#build",
         "@batijs/tailwindcss#build",
         "@batijs/telefunc#build",
@@ -47,7 +50,8 @@
         "@batijs/vercel#build",
         "@batijs/vue#build",
         "@batijs/vue-firebase-auth#build",
-        "@batijs/vue-lucia-auth#build"
+        "@batijs/vue-lucia-auth#build",
+        "@batijs/vue-sentry#build"
       ],
       "outputs": [
         "dist/**"
